Restructure tests according to file renaming
